Deleted unneccesary files
